+++ multiqc/modules/samtools/flagstat.py
@@ -37,13 +37,23 @@ class FlagstatReportMixin():
 
             # General Stats Table
             flagstats_headers = dict()
+            flagstats_headers['flagstat_total'] = {
+                'title': '{} Total Reads'.format(config.read_count_prefix),
+                'description': 'Total Mapped in the bam file ({})'.format(config.read_count_desc),
+                'min': 0,
+                'modify': lambda x: x * config.read_count_multiplier,
+                'shared_key': 'read_count',
+                'placement' : 100.0,
+                'hidden': True
+
+            }
             flagstats_headers['mapped_passed'] = {
                 'title': '{} Reads Mapped'.format(config.read_count_prefix),
                 'description': 'Reads Mapped in the bam file ({})'.format(config.read_count_desc),
                 'min': 0,
                 'modify': lambda x: x * config.read_count_multiplier,
                 'shared_key': 'read_count',
-                'placement' : 100.0
+                'placement' : 101.0
             }
             self.general_stats_addcols(self.samtools_flagstat, flagstats_headers)